@@ -6,10 +6,13 @@
     "@cucumber/cucumber": "^11.2.0",
     "@eslint/js": "^9.20.0",
     "@types/cucumber": "^6.0.1",
-    "@types/node": "^14.0.27",
+    "@types/node": "^14.18.63",
+    "@typescript-eslint/eslint-plugin": "^8.24.0",
+    "@typescript-eslint/parser": "^8.24.0",
     "eslint": "^9.20.0",
     "globals": "^15.14.0",
-    "typescript": "^4.5.4"
+    "ts-node": "^10.9.2",
+    "typescript": "^4.9.5"
   },
   "scripts": {
     "test": "cucumber-js --profile default",
