@@ -1,19 +1,19 @@
 const { Command } = require('commander');
-const repository = require('../Infra/Repository');
-const Fleet = require('../Domain/Fleet');
-const Vehicle = require('../Domain/Vehicle');
-const Location = require('../Domain/Location');
+import * as repository from '../Infra/Repository';
+import { Fleet } from '../Domain/Fleet';
+import { Vehicle } from '../Domain/Vehicle';
+import { Location } from '../Domain/Location';
 
 const app = new Command()
 
 app
     .command("create <userId>")
     .description("Create a fleet for a user")
-    .action((userId) => {
+    .action((userId: string) => {
         try {
             const fleet = new Fleet(userId);
             repository.createFleet(fleet);
-        } catch (error) {
+        } catch (error: any) {
             console.log(error.message);
         }
     });
@@ -21,11 +21,11 @@ app
 app
     .command("register-vehicle <fleetId> <vehiclePlateNumber>")
     .description("Register a vehicle in a fleet")
-    .action((fleetId, vehiclePlateNumber) => {
+    .action((fleetId: string, vehiclePlateNumber: string) => {
         try {
             const vehicle = new Vehicle(vehiclePlateNumber);
             repository.registerVehicle(fleetId, vehicle);
-        } catch (error) {
+        } catch (error: any) {
             console.log(error.message);
         }
     });
@@ -33,12 +33,12 @@ app
 app
     .command("localize-vehicle <fleetId> <vehiclePlateNumber> <lat> <lng> [alt]")
     .description("Localize a vehicle to a location")
-    .action((fleetId, vehiclePlateNumber, lat, lng, alt) => {
-        alt = alt || 0;
+    .action((fleetId: string, vehiclePlateNumber: string, lat: number, lng: number, alt: number = 0) => {
+        //alt = alt || 0;
         try {
             const location = new Location(lat, lng, alt);
             repository.localizeVehicle(fleetId, vehiclePlateNumber, location);
-        } catch (error) {
+        } catch (error: any) {
             console.log(error.message);
         }
     });
